Create ServiceBridgeSupervisor.scala to allow for better actor supervision within the application.
